scripts and NCL scripts for running Forecast Sensitivity to Observations. They are not fully tested yet, but they are committed now in order for easier future development. Expecting more changes from Tom Auligne.
M    scripts/da_run_suite.ksh
AM   scripts/da_run_sensitivity.ksh
M    scripts/da_set_defaults.ksh
M    scripts/da_run_wrf.ksh
M    scripts/da_run_wrfvar.ksh
AM   graphics/ncl/ini_target_grad4.ncl
AM   graphics/ncl/Diag1_target_measure.ncl
A    graphics/ncl/wrf_sum.ncl
A    graphics/ncl/Diag2_target_measure.ncl


git-svn-id: https://svn-wrf-var.cgd.ucar.edu/tools@3982 70815b0f-b814-0410-81f7-dfa10a885da6

+++ WRFDA/var/scripts/da_run_wrf.ksh
@@ -10,14 +10,18 @@
 # think it necessary then please email wrfhelp@ucar.edu with details.
 #########################################################################
 
+# Option to run WRF, WRFNL, WRFTL and WRFAD
+if [[ $# -gt 0 ]]; then WRF_CONF=$1; else WRF_CONF=""; fi
+echo "Running WRF Configuration: WRF"$WRF_CONF
+
 export REL_DIR=${REL_DIR:-$HOME/trunk}
-export WRFVAR_DIR=${WRFVAR_DIR:-$REL_DIR/WRFDA}
+export WRF_DIR=${WRF_DIR:-$REL_DIR/WRFNL}
 export SCRIPTS_DIR=${SCRIPTS_DIR:-$WRFVAR_DIR/var/scripts}
 . ${SCRIPTS_DIR}/da_set_defaults.ksh
 export RUN_DIR=${RUN_DIR:-$EXP_DIR/wrf}
 export WORK_DIR=$RUN_DIR/working
-
 export NL_RUN_HOURS=$FCST_RANGE
+#let NL_RUN_SECONDS=$FCST_RANGE*3600
 
 # allow for ensemble members identified by CMEM
 export FC_DIR_DATE=$FC_DIR/$DATE
@@ -26,9 +30,83 @@ if [[ ! -z $CMEM ]]; then
 fi
 
 if [[ ! -d $FC_DIR_DATE ]]; then mkdir -p $FC_DIR_DATE; fi
-rm -rf $WORK_DIR
+if [[ $WRF_CONF == "" ]]; then rm -rf $WORK_DIR; fi
 mkdir -p $RUN_DIR $WORK_DIR
 cd $WORK_DIR
+
+#----- WRF -----
+if [[ $WRF_CONF == "" ]]; then
+   export EXEC_DIR=$WRF_DIR
+   export EXEC_FILE="wrf.exe"   
+else
+   export NL_INPUTOUT_BEGIN_H=0
+   export NL_INPUTOUT_END_H=$FCST_RANGE
+   export NL_INPUTOUT_INTERVAL=60 
+   export NL_HISTORY_INTERVAL=60
+   export NL_FRAMES_PER_AUXHIST2=1
+   
+   let TIME_STEP_MINUTE=$NL_TIME_STEP/60
+   if [[ -f $WORK_DIR/namelist.input ]];  then rm -rf $WORK_DIR/namelist.input;  fi
+   if [[ -f $WORK_DIR/namelist.output ]]; then rm -rf $WORK_DIR/namelist.output; fi
+   
+#----- WRFNL -----
+   if   [[ $WRF_CONF == "NL" ]]; then
+      export EXEC_DIR=$WRFNL_DIR
+      export EXEC_FILE="wrf.exe"
+      export NL_INPUT_OUTNAME="./nl_d<domain>_<date>"
+      export NL_WRITE_INPUT=false
+      export NL_AUXHIST2_OUTNAME="./auxhist2_d<domain>_<date>"
+      export NL_AUXHIST2_BEGIN_H=0
+      export NL_AUXHIST2_END_H=$FCST_RANGE
+      export NL_AUXHIST2_INTERVAL=$TIME_STEP_MINUTE
+      export NL_IO_FORM_AUXHIST2=2
+
+#----- WRFPLUS -----
+   else
+      export EXEC_DIR=$WRFPLUS_DIR
+      export EXEC_FILE="wrfplus.exe"
+      if [[ $WRF_CONF == "TL" ]]; then 
+         export NL_INPUT_OUTNAME="./tl_d<domain>_<date>"
+         export NL_DYN_OPT=202
+      elif [[ $WRF_CONF == "AD" ]]; then 
+         export NL_INPUT_OUTNAME="./ad_d<domain>_<date>"
+         export NL_DYN_OPT=302
+      fi	
+      if [[ $WRF_CONF == "AD" ]]; then 
+         export NL_AUXINPUT3_INNAME="./auxinput3_d<domain>_<date>"
+         let NL_AUXINPUT3_INTERVAL_S=$FCST_RANGE*3600      
+         export NL_INPUTOUT_INTERVAL_S=3600 
+         export NL_AUXINPUT2_BEGIN_S=0 
+         let NL_AUXINPUT2_END_S=$FCST_RANGE*3600
+         export NL_INTERVAL_SECONDS=3600
+         export NL_IO_FORM_AUXINPUT3=2
+      else
+         export NL_INPUTOUT_INTERVAL=60	 
+      fi
+      export NL_AUXINPUT2_INNAME="./auxhist2_d<domain>_<date>"
+      export NL_AUXINPUT2_INTERVAL=$TIME_STEP_MINUTE
+      export NL_IO_FORM_AUXINPUT2=2
+
+      export NL_NPROC_X=0            # Parallelization
+      export NL_MP_PHYSICS=0         # Physics
+      export NL_RA_LW_PHYSICS=0
+      export NL_RA_SW_PHYSICS=0
+      export NL_RADT=0
+      export NL_SF_SFCLAY_PHYSICS=0
+      export NL_SF_SURFACE_PHYSICS=0
+      export NL_BL_PBL_PHYSICS=0
+      export NL_CU_PHYSICS=0
+      export NL_CUDT=0
+      export NL_W_DAMPING=0          # Damping for Convective Updraft
+      export NL_DIFF_OPT=0           # Diffusion (horiz + vert)
+      export NL_DAMPCOEF=0.01
+      export NL_ISFFLX=0             # Surface fluxes
+      export NL_JCDFI_IO=false
+      export NL_TIME_STEP_SOUND=4    # Nb of sound time-steps per RK step: 0->auto
+      
+      if [[ ! -d trace ]]; then mkdir trace; fi    
+   fi
+fi
 
 # Get extra namelist variables:
 . $SCRIPTS_DIR/da_get_date_range.ksh
@@ -39,7 +117,7 @@ echo "<H1>$EXPT wrf</H1><PRE>"
 date
 
 echo 'REL_DIR        <A HREF="file:'$REL_DIR'">'$REL_DIR'</a>'         
-echo 'WRF_DIR        <A HREF="file:'$WRF_DIR'">'$WRF_DIR'</a>' $WRF_VN 
+echo 'WRF'${WRF_CONF}'_DIR      <A HREF="file:'$EXEC_DIR'">'$EXEC_DIR'</a>' $WRF_VN 
 echo 'RUN_DIR        <A HREF="file:'$RUN_DIR'">'$RUN_DIR'</a>'         
 echo 'WORK_DIR       <A HREF="file:'$WORK_DIR'">'$WORK_DIR'</a>'       
 echo 'RC_DIR         <A HREF="file:'$RC_DIR'">'$RC_DIR'</a>'           
@@ -53,20 +131,19 @@ echo "DOMAINS        $DOMAINS"
 echo "MEM            $MEM"
 
 # Copy necessary info (better than link as not overwritten):
-ln -fs $WRF_DIR/main/wrf.exe .
-ln -fs $WRF_DIR/run/gribmap.txt .
-ln -fs $WRF_DIR/run/ozone* .
-ln -fs $WRF_DIR/run/*.TBL .
-ln -fs $WRF_DIR/run/*.tbl .
+ln -fs $EXEC_DIR/main/$EXEC_FILE .
+ln -fs $EXEC_DIR/run/gribmap.txt .
+#ln -fs $EXEC_DIR/run/ozone* .
+ln -fs $EXEC_DIR/run/*.TBL .
 if $DOUBLE; then
-   ln -fs $WRF_DIR/run/RRTM_DATA_DBL RRTM_DATA
-   ln -fs $WRF_DIR/run/ETAMPNEW_DATA_DBL ETAMPNEW_DATA
+   ln -fs $EXEC_DIR/run/RRTM_DATA_DBL RRTM_DATA
+   ln -fs $EXEC_DIR/run/ETAMPNEW_DATA_DBL ETAMPNEW_DATA
 else
-   ln -fs $WRF_DIR/run/RRTM_DATA .
-   ln -fs $WRF_DIR/run/ETAMPNEW_DATA .
+   ln -fs $EXEC_DIR/run/RRTM_DATA .
+   ln -fs $EXEC_DIR/run/ETAMPNEW_DATA .
 fi
-ln -fs $WRF_DIR/run/CAM_ABS_DATA .
-ln -fs $WRF_DIR/run/CAM_AEROPT_DATA .
+ln -fs $EXEC_DIR/run/CAM_ABS_DATA .
+ln -fs $EXEC_DIR/run/CAM_AEROPT_DATA .
 
 for DOMAIN in $DOMAINS; do
    # Copy this file, so the copy back of wrfinput files later does
@@ -75,20 +152,20 @@ for DOMAIN in $DOMAINS; do
    # WHY
    # cp ${RC_DIR}/$DATE/wrflowinp_d${DOMAIN} wrflowinp_d${DOMAIN}
 done
-ln -fs $WRF_INPUT_DIR/wrfbdy_d01 .
+cp $WRF_INPUT_DIR/wrfbdy_d01* .
 
 let NL_INTERVAL_SECONDS=$LBC_FREQ*3600
 
 if [[ $WRF_NAMELIST'.' != '.' ]]; then
    ln -fs $WRF_NAMELIST namelist.input
-elif [[ -f $WRF_DIR/inc/namelist_script.inc ]]; then
-   . $WRF_DIR/inc/namelist_script.inc
+elif [[ -f $EXEC_DIR/inc/namelist_script.inc ]]; then
+   . $EXEC_DIR/inc/namelist_script.inc "namelist.input"
 else
-   ln -fs $WRF_DIR/test/em_real/namelist.input .
+   ln -fs $EXEC_DIR/test/em_real/namelist.input .
 fi
 
-cp namelist.input $RUN_DIR
-
+cp namelist.input $RUN_DIR/namelist.input
+cp namelist.input $RUN_DIR/namelist.input$WRF_CONF
 echo '<A HREF="namelist.input">Namelist input</a>'
 
 if $DUMMY; then
@@ -106,23 +183,26 @@ if $DUMMY; then
       LOCAL_DATE=$($BUILD_DIR/da_advance_time.exe $LOCAL_DATE $NL_HISTORY_INTERVAL)
    done
 else
-   if $NL_VAR4D && [[ $NUM_PROCS -gt 1 ]]; then
-      touch wrfnl_go_ahead
-   fi
-   $RUN_CMD ./wrf.exe
+   if $NL_VAR4D && [[ $NUM_PROCS -gt 1 ]]; then touch wrfnl_go_ahead; fi
+   if $RUN_ADJ_SENS && [[ $NUM_PROCS -gt 1 ]]; then touch wrfnl_go_ahead; touch wrf_go_ahead; fi
+   if $RUN_TL_TEST && [[ $NUM_PROCS -gt 1 ]];  then touch wrfnl_go_ahead; touch wrf_go_ahead; fi
+
+   $RUN_CMD $EXEC_FILE
+   
    if [[ -f rsl.out.0000 ]]; then
-      grep -q 'SUCCESS COMPLETE WRF' rsl.out.0000 
+      grep -q 'SUCCESS COMPLETE ' $EXEC_FILE rsl.out.0000 
       RC=$?
    fi
 
-   cp namelist.output $RUN_DIR
+   cp namelist.output $RUN_DIR/namelist.output
+   cp namelist.output $RUN_DIR/namelist.output$WRF_CONF
    echo '<A HREF="namelist.output">Namelist output</a>'
 
    if [[ -f rsl.out.0000 ]]; then
-      rm -rf $RUN_DIR/rsl
-      mkdir -p $RUN_DIR/rsl
-      mv rsl* $RUN_DIR/rsl
-      cd $RUN_DIR/rsl
+      rm -rf $RUN_DIR/rsl_$WRF_CONF
+      mkdir -p $RUN_DIR/rsl_$WRF_CONF
+      mv rsl* $RUN_DIR/rsl_$WRF_CONF
+      cd $RUN_DIR/rsl_$WRF_CONF
       for FILE in rsl*; do
          echo "<HTML><HEAD><TITLE>$FILE</TITLE></HEAD>" > $FILE.html
          echo "<H1>$FILE</H1><PRE>" >> $FILE.html
@@ -130,20 +210,18 @@ else
          echo "</PRE></BODY></HTML>" >> $FILE.html
          rm $FILE
       done
-      echo '<A HREF="rsl/rsl.out.0000.html">rsl.out.0000</a>'
-      echo '<A HREF="rsl/rsl.error.0000.html">rsl.error.0000</a>'
-      echo '<A HREF="rsl">Other RSL output</a>'
+      echo '<A HREF="rsl_'${WRF_CONF}'/rsl.out.0000.html">rsl.out.0000</a>'
+      echo '<A HREF="rsl_'${WRF_CONF}'/rsl.error.0000.html">rsl.error.0000</a>'
+      echo '<A HREF="rsl_'${WRF_CONF}'>Other RSL output</a>'
    fi
-#  mv wrfout* wrfinput* $FC_DIR/$DATE
-# else
-#    echo "$FC_DIR/$DATE/wrfout_d${DOMAIN}_${END_YEAR}-${END_MONTH}-${END_DAY}_${END_HOUR}:00:00 already exists, skipping"
-# fi
 
    echo $(date +'%D %T') "Ended $RC"
 fi
 
-mv $WORK_DIR/wrfinput_* $FC_DIR_DATE
-mv $WORK_DIR/wrfout_* $FC_DIR_DATE
+if [[ $WRF_CONF == "" ]]; then
+   mv $WORK_DIR/wrfinput_* $FC_DIR_DATE
+   mv $WORK_DIR/wrfout_*   $FC_DIR_DATE
+fi
 
 if $CLEAN; then
    rm -rf $WORK_DIR
